Refinements
Noticed in the Chrome resources tab that blindly changing the main
image's src attribute would cause the browser to send GET requests.
Need to fix this:
- instead of .main-img being only the first image, dump ALL product
images as .main-img
- remove the .preload-sku-img div and dump sku images as .main-img too
- refactor main image change logic into new changeImage function

Minor misc refactor
- move some JS constants to the top
- split paginate function into paginate (event handler) and changePage
> also added a return value to changePage to make life easier with
changeSku

Noticed a WTF flicker when user changes selected thumbnail and then
hovers over different sku colours.  Reduce flicker effect by emulating
main qvc idp.

Original logic is to reset to the previous selected image when a user
mouses out of a sku option.  If the user mouses out of one and
immediately mouses over another sku, this presents itself as a wtf
flicker effect.

QVC seems to try to get around this issue by immediately resetting the
chosen thumbnail to the first one (the "primary image" for the item).
It appears that sku images look almost identical to the first image
(except for the colour), so although there is still flickering, the near
identical-ness makes it less obvious.