Prelim basic keyboard support
The $thumbnails.length check is to ensure we're attaching event
listeners when we need to.  Noticed that when the gallery is loaded in a
lightbox on the homepage, the first keypress will move the thumbnail
selection by 2.  Traced it down to main.js running
$('.gallery-ujs').gallery() which causes init to run on ... nothing.
But since init is still run, the event listener is added and when the
quickview is opened, we effectively attached two event listeners =
keypress moves thumbnail by 2.  So this is basically a hack for this
case.

Also added a minor UI tweak.  We have enough space to show 7.5 images in
the thumbnail area (since we have to take in account the pagination
buttons).  But if a product has exactly 8 images, we have enough space
to show them all, so treat a product having 8 images as a show_buttons =
false case.

TODO: need keyboard to paginate ... but doesn't seem like any products
have more than 8 images, so this feature/bugfix doesn't seem so high
priority due to above enhancement